requirements.txt
Flask==3.0.3
twilio==9.2.3
gunicorn==23.0.0
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
python-dotenv==1.0.0
